requirements.txt
opencv-python
langchain
langchain-community
langchain-openai
langchain_google_genai
openai
python-dotenv
pyaudio
soundfile
SpeechRecognition
git+https://github.com/openai/whisper.git
Pillow
numpy